// 10/src/types.ts
export type TokenType =
  | "KEYWORD"
  | "SYMBOL"
  | "IDENTIFIER"
  | "INT_CONST"
  | "STRING_CONST";

export type Keyword =
  | "CLASS"
  | "METHOD"
  | "FUNCTION"
  | "CONSTRUCTOR"
  | "INT"
  | "BOOLEAN"
  | "CHAR"
  | "VOID"
  | "VAR"
  | "STATIC"
  | "FIELD"
  | "LET"
  | "DO"
  | "IF"
  | "ELSE"
  | "WHILE"
  | "RETURN"
  | "TRUE"
  | "FALSE"
  | "NULL"
  | "THIS";